Update dependencies and refactor code for YOLO conversion (#73)
- Updated version in Cargo.lock and Cargo.toml to 0.2.7.
- Added new dependencies: `autocfg`, `dashmap`, `hashbrown`, `lock_api`, `once_cell`, `parking_lot_core`, `redox_syscall`, and `scopeguard`.
- Introduced new modules for configuration, dataset processing, and I/O operations.
- Refactored main.rs to streamline the dataset processing pipeline.
- Enhanced JSON parsing to handle large files efficiently.
- Improved command-line argument parsing for better usability.
