+++ Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "labelme2yolo"
-version = "0.2.6"
+version = "0.2.7"
 edition = "2021"
 
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
@@ -8,12 +8,13 @@ edition = "2021"
 [dependencies]
 clap = { version = "4.5.11", features = ["derive"] }
 glob = "0.3.1"
-serde = { version = "1.0", features = ["derive"] }  
+serde = { version = "1.0", features = ["derive"] }
 serde_json = "1.0"
 rand = "0.8"
 sanitize-filename = "0.5"
 indicatif = "0.16"
 rayon = "1.5"
+dashmap = "5.5"
 base64 = "0.13"
 log = "0.4"
 env_logger = "0.9"
