+++ pyproject.toml
@@ -9,13 +9,11 @@ readme = "README.md"
 requires-python = ">=3.8"
 license = "MIT"
 keywords = []
-authors = [
-  { name = "GreatV(Wang Xin)", email = "xinwang614@gmail.com" },
-]
+authors = [{ name = "GreatV(Wang Xin)", email = "xinwang614@gmail.com" }]
 classifiers = [
-    "Programming Language :: Rust",
-    "Programming Language :: Python :: Implementation :: CPython",
-    "Programming Language :: Python :: Implementation :: PyPy",
+  "Programming Language :: Rust",
+  "Programming Language :: Python :: Implementation :: CPython",
+  "Programming Language :: Python :: Implementation :: PyPy",
 ]
 
 dynamic = ["version"]
